+++ src/main/java/finalmission/dto/response/MovieSlotCreateResponse.java
@@ -3,5 +3,5 @@ package finalmission.dto.response;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public record MovieSlotCreateResponse(Long id, String movieName, LocalDate date, LocalTime startAt, Long seats) {
+public record MovieSlotCreateResponse(Long id, String movieName, LocalDate date, LocalTime startAt, Integer seats) {
 }
